fix(grids): move directives next to modules to ensure compiled order
