+++ projects/igniteui-angular/grids/grid/src/grid.module.ts
@@ -1,5 +1,190 @@
 import { NgModule } from '@angular/core';
-import { IGX_GRID_DIRECTIVES } from './public_api';
+import { IgxGridComponent } from './grid.component';
+
+/* Imports that cannot be resolved from IGX_GRID_COMMON_DIRECTIVES spread
+    NOTE: Do not remove! Issue: https://github.com/IgniteUI/igniteui-angular/issues/13310
+*/
+import {
+    IgxRowDirective,
+    IgxGridFooterComponent,
+    IgxAdvancedFilteringDialogComponent,
+    IgxHeaderCollapsedIndicatorDirective,
+    IgxHeaderExpandedIndicatorDirective,
+    IgxRowCollapsedIndicatorDirective,
+    IgxRowExpandedIndicatorDirective,
+    IgxSortAscendingHeaderIconDirective,
+    IgxSortDescendingHeaderIconDirective,
+    IgxSortHeaderIconDirective,
+    IgxGridEmptyTemplateDirective,
+    IgxGridLoadingTemplateDirective,
+    IgxExcelStyleHeaderIconDirective,
+    IgxDragIndicatorIconDirective,
+    IgxRowDragGhostDirective,
+    IgxGridStateDirective,
+    IgxGridHeaderComponent,
+    IgxGridHeaderGroupComponent,
+    IgxGridHeaderRowComponent,
+    IgxFilterCellTemplateDirective,
+    IgxSummaryTemplateDirective,
+    IgxCellTemplateDirective,
+    IgxCellValidationErrorDirective,
+    IgxCellHeaderTemplateDirective,
+    IgxCellFooterTemplateDirective,
+    IgxCellEditorTemplateDirective,
+    IgxCollapsibleIndicatorTemplateDirective,
+    IgxColumnComponent,
+    IgxColumnGroupComponent,
+    IgxColumnLayoutComponent,
+    IgxColumnRequiredValidatorDirective,
+    IgxColumnMinValidatorDirective,
+    IgxColumnMaxValidatorDirective,
+    IgxColumnEmailValidatorDirective,
+    IgxColumnMinLengthValidatorDirective,
+    IgxColumnMaxLengthValidatorDirective,
+    IgxColumnPatternValidatorDirective,
+    IgxColumnActionsComponent,
+    IgxColumnHidingDirective,
+    IgxColumnPinningDirective,
+    IgxRowSelectorDirective,
+    IgxGroupByRowSelectorDirective,
+    IgxHeadSelectorDirective,
+    IgxCSVTextDirective,
+    IgxExcelTextDirective,
+    IgxGridToolbarActionsComponent,
+    IgxGridToolbarAdvancedFilteringComponent,
+    IgxGridToolbarComponent,
+    IgxGridToolbarExporterComponent,
+    IgxGridToolbarHidingComponent,
+    IgxGridToolbarPinningComponent,
+    IgxGridToolbarTitleComponent,
+    IgxGridToolbarDirective,
+    IgxGridExcelStyleFilteringComponent,
+    IgxExcelStyleHeaderComponent,
+    IgxExcelStyleSortingComponent,
+    IgxExcelStylePinningComponent,
+    IgxExcelStyleHidingComponent,
+    IgxExcelStyleSelectingComponent,
+    IgxExcelStyleClearFiltersComponent,
+    IgxExcelStyleConditionalFilterComponent,
+    IgxExcelStyleMovingComponent,
+    IgxExcelStyleSearchComponent,
+    IgxExcelStyleColumnOperationsTemplateDirective,
+    IgxExcelStyleFilterOperationsTemplateDirective,
+    IgxExcelStyleLoadingValuesTemplateDirective,
+    IgxGridDetailTemplateDirective,
+    IgxGroupByRowTemplateDirective,
+    IgxRowAddTextDirective,
+    IgxRowEditActionsDirective,
+    IgxRowEditTabStopDirective,
+    IgxRowEditTextDirective,
+    IgxGridActionButtonComponent,
+    IgxGridPinningActionsComponent,
+    IgxGridActionsBaseDirective,
+    IgxGridEditingActionsComponent
+} from "igniteui-angular/grids/core";
+import {
+    IgxPaginatorComponent,
+    IgxPageNavigationComponent,
+    IgxPageSizeSelectorComponent,
+    IgxPaginatorContentDirective,
+    IgxPaginatorDirective
+} from 'igniteui-angular/paginator';
+
+/* NOTE: Grid directives collection for ease-of-use import in standalone components scenario */
+export const IGX_GRID_DIRECTIVES = [
+    IgxGridComponent,
+    IgxGroupByRowTemplateDirective,
+    IgxGridDetailTemplateDirective,
+    IgxRowAddTextDirective,
+    IgxRowEditActionsDirective,
+    IgxRowEditTextDirective,
+    IgxRowEditTabStopDirective,
+    // IGX_GRID_COMMON_DIRECTIVES:
+    IgxRowDirective,
+    IgxGridFooterComponent,
+    IgxAdvancedFilteringDialogComponent,
+    IgxRowExpandedIndicatorDirective,
+    IgxRowCollapsedIndicatorDirective,
+    IgxHeaderExpandedIndicatorDirective,
+    IgxHeaderCollapsedIndicatorDirective,
+    IgxExcelStyleHeaderIconDirective,
+    IgxSortAscendingHeaderIconDirective,
+    IgxSortDescendingHeaderIconDirective,
+    IgxSortHeaderIconDirective,
+    IgxGridEmptyTemplateDirective,
+    IgxGridLoadingTemplateDirective,
+    IgxDragIndicatorIconDirective,
+    IgxRowDragGhostDirective,
+    IgxGridStateDirective,
+    // IGX_GRID_ACTIONS
+    IgxGridPinningActionsComponent,
+    IgxGridEditingActionsComponent,
+    IgxGridActionsBaseDirective,
+    IgxGridActionButtonComponent,
+    // IGX_GRID_HEADERS_DIRECTIVES:
+    IgxGridHeaderComponent,
+    IgxGridHeaderGroupComponent,
+    IgxGridHeaderRowComponent,
+    // IGX_GRID_COLUMN_DIRECTIVES:
+    IgxFilterCellTemplateDirective,
+    IgxSummaryTemplateDirective,
+    IgxCellTemplateDirective,
+    IgxCellValidationErrorDirective,
+    IgxCellHeaderTemplateDirective,
+    IgxCellFooterTemplateDirective,
+    IgxCellEditorTemplateDirective,
+    IgxCollapsibleIndicatorTemplateDirective,
+    IgxColumnComponent,
+    IgxColumnGroupComponent,
+    IgxColumnLayoutComponent,
+    // IGX_GRID_COLUMN_ACTIONS_DIRECTIVES:
+    IgxColumnActionsComponent,
+    IgxColumnHidingDirective,
+    IgxColumnPinningDirective,
+    // IGX_GRID_SELECTION_DIRECTIVES:
+    IgxRowSelectorDirective,
+    IgxGroupByRowSelectorDirective,
+    IgxHeadSelectorDirective,
+    // IGX_GRID_TOOLBAR_DIRECTIVES:
+    IgxCSVTextDirective,
+    IgxExcelTextDirective,
+    IgxGridToolbarActionsComponent,
+    IgxGridToolbarAdvancedFilteringComponent,
+    IgxGridToolbarComponent,
+    IgxGridToolbarExporterComponent,
+    IgxGridToolbarHidingComponent,
+    IgxGridToolbarPinningComponent,
+    IgxGridToolbarTitleComponent,
+    IgxGridToolbarDirective,
+    // IGX_GRID_EXCEL_STYLE_FILTER_DIRECTIVES:
+    IgxGridExcelStyleFilteringComponent,
+    IgxExcelStyleHeaderComponent,
+    IgxExcelStyleSortingComponent,
+    IgxExcelStylePinningComponent,
+    IgxExcelStyleHidingComponent,
+    IgxExcelStyleSelectingComponent,
+    IgxExcelStyleClearFiltersComponent,
+    IgxExcelStyleConditionalFilterComponent,
+    IgxExcelStyleMovingComponent,
+    IgxExcelStyleSearchComponent,
+    IgxExcelStyleColumnOperationsTemplateDirective,
+    IgxExcelStyleFilterOperationsTemplateDirective,
+    IgxExcelStyleLoadingValuesTemplateDirective,
+    // IGX_GRID_VALIDATION_DIRECTIVES:
+    IgxColumnRequiredValidatorDirective,
+    IgxColumnMinValidatorDirective,
+    IgxColumnMaxValidatorDirective,
+    IgxColumnEmailValidatorDirective,
+    IgxColumnMinLengthValidatorDirective,
+    IgxColumnMaxLengthValidatorDirective,
+    IgxColumnPatternValidatorDirective,
+    // IGX_PAGINATOR_DIRECTIVES:
+    IgxPaginatorComponent,
+    IgxPageNavigationComponent,
+    IgxPageSizeSelectorComponent,
+    IgxPaginatorContentDirective,
+    IgxPaginatorDirective
+] as const;
 
 /**
  * @hidden
