+++ projects/igniteui-angular/grids/hierarchical-grid/src/public_api.ts
@@ -1,188 +1,5 @@
-import { IgxHierarchicalGridComponent } from './hierarchical-grid.component';
-import { IgxRowIslandComponent } from './row-island.component';
-
 export * from './events';
 export * from './hierarchical-grid.component';
 export * from './row-island.component';
 export * from './hierarchical-grid.pipes';
-
-/* Imports that cannot be resolved from IGX_GRID_COMMON_DIRECTIVES spread
-    NOTE: Do not remove! Issue: https://github.com/IgniteUI/igniteui-angular/issues/13310
-*/
-/* eslint-disable @typescript-eslint/no-unused-vars */
-import {
-    IgxRowDirective,
-    IgxRowEditTextDirective,
-    IgxRowAddTextDirective,
-    IgxRowEditActionsDirective,
-    IgxRowEditTabStopDirective,
-    IgxGridFooterComponent,
-    IgxAdvancedFilteringDialogComponent,
-    IgxHeaderCollapsedIndicatorDirective,
-    IgxHeaderExpandedIndicatorDirective,
-    IgxRowCollapsedIndicatorDirective,
-    IgxRowExpandedIndicatorDirective,
-    IgxSortAscendingHeaderIconDirective,
-    IgxSortDescendingHeaderIconDirective,
-    IgxSortHeaderIconDirective,
-    IgxExcelStyleHeaderIconDirective,
-    IgxDragIndicatorIconDirective,
-    IgxRowDragGhostDirective,
-    IgxGridStateDirective,
-    IgxGridHeaderComponent,
-    IgxGridHeaderGroupComponent,
-    IgxGridHeaderRowComponent,
-    IgxFilterCellTemplateDirective,
-    IgxSummaryTemplateDirective,
-    IgxCellTemplateDirective,
-    IgxCellValidationErrorDirective,
-    IgxCellHeaderTemplateDirective,
-    IgxCellFooterTemplateDirective,
-    IgxCellEditorTemplateDirective,
-    IgxCollapsibleIndicatorTemplateDirective,
-    IgxColumnComponent,
-    IgxColumnGroupComponent,
-    IgxColumnLayoutComponent,
-    IgxColumnRequiredValidatorDirective,
-    IgxColumnMinValidatorDirective,
-    IgxColumnMaxValidatorDirective,
-    IgxColumnEmailValidatorDirective,
-    IgxColumnMinLengthValidatorDirective,
-    IgxColumnMaxLengthValidatorDirective,
-    IgxColumnPatternValidatorDirective,
-    IgxColumnActionsComponent,
-    IgxColumnHidingDirective,
-    IgxColumnPinningDirective,
-    IgxRowSelectorDirective,
-    IgxGroupByRowSelectorDirective,
-    IgxHeadSelectorDirective,
-    IgxCSVTextDirective,
-    IgxExcelTextDirective,
-    IgxGridToolbarActionsComponent,
-    IgxGridToolbarAdvancedFilteringComponent,
-    IgxGridToolbarComponent,
-    IgxGridToolbarExporterComponent,
-    IgxGridToolbarHidingComponent,
-    IgxGridToolbarPinningComponent,
-    IgxGridToolbarTitleComponent,
-    IgxGridToolbarDirective,
-    IgxGridExcelStyleFilteringComponent,
-    IgxExcelStyleHeaderComponent,
-    IgxExcelStyleSortingComponent,
-    IgxExcelStylePinningComponent,
-    IgxExcelStyleHidingComponent,
-    IgxExcelStyleSelectingComponent,
-    IgxExcelStyleClearFiltersComponent,
-    IgxExcelStyleConditionalFilterComponent,
-    IgxExcelStyleMovingComponent,
-    IgxExcelStyleSearchComponent,
-    IgxExcelStyleColumnOperationsTemplateDirective,
-    IgxExcelStyleFilterOperationsTemplateDirective,
-    IgxExcelStyleLoadingValuesTemplateDirective,
-    IgxGridActionButtonComponent,
-    IgxGridActionsBaseDirective,
-    IgxGridEditingActionsComponent,
-    IgxGridPinningActionsComponent
-} from "igniteui-angular/grids/core";
-import {
-    IgxPaginatorComponent,
-    IgxPageNavigationComponent,
-    IgxPageSizeSelectorComponent,
-    IgxPaginatorContentDirective,
-    IgxPaginatorDirective
-} from 'igniteui-angular/paginator';
-
-/* NOTE: Hierarchical grid directives collection for ease-of-use import in standalone components scenario */
-export const IGX_HIERARCHICAL_GRID_DIRECTIVES = [
-    IgxHierarchicalGridComponent,
-    IgxRowIslandComponent,
-    IgxRowAddTextDirective,
-    IgxRowEditActionsDirective,
-    IgxRowEditTextDirective,
-    IgxRowEditTabStopDirective,
-    // IGX_GRID_COMMON_DIRECTIVES:
-    IgxRowDirective,
-    IgxGridFooterComponent,
-    IgxAdvancedFilteringDialogComponent,
-    IgxRowExpandedIndicatorDirective,
-    IgxRowCollapsedIndicatorDirective,
-    IgxHeaderExpandedIndicatorDirective,
-    IgxHeaderCollapsedIndicatorDirective,
-    IgxExcelStyleHeaderIconDirective,
-    IgxSortAscendingHeaderIconDirective,
-    IgxSortDescendingHeaderIconDirective,
-    IgxSortHeaderIconDirective,
-    IgxDragIndicatorIconDirective,
-    IgxRowDragGhostDirective,
-    IgxGridStateDirective,
-    // IGX_GRID_ACTIONS
-    IgxGridPinningActionsComponent,
-    IgxGridEditingActionsComponent,
-    IgxGridActionsBaseDirective,
-    IgxGridActionButtonComponent,
-    // IGX_GRID_HEADERS_DIRECTIVES:
-    IgxGridHeaderComponent,
-    IgxGridHeaderGroupComponent,
-    IgxGridHeaderRowComponent,
-    // IGX_GRID_COLUMN_DIRECTIVES:
-    IgxFilterCellTemplateDirective,
-    IgxSummaryTemplateDirective,
-    IgxCellTemplateDirective,
-    IgxCellValidationErrorDirective,
-    IgxCellHeaderTemplateDirective,
-    IgxCellFooterTemplateDirective,
-    IgxCellEditorTemplateDirective,
-    IgxCollapsibleIndicatorTemplateDirective,
-    IgxColumnComponent,
-    IgxColumnGroupComponent,
-    IgxColumnLayoutComponent,
-    // IGX_GRID_COLUMN_ACTIONS_DIRECTIVES:
-    IgxColumnActionsComponent,
-    IgxColumnHidingDirective,
-    IgxColumnPinningDirective,
-    // IGX_GRID_SELECTION_DIRECTIVES:
-    IgxRowSelectorDirective,
-    IgxGroupByRowSelectorDirective,
-    IgxHeadSelectorDirective,
-    // IGX_GRID_TOOLBAR_DIRECTIVES:
-    IgxCSVTextDirective,
-    IgxExcelTextDirective,
-    IgxGridToolbarActionsComponent,
-    IgxGridToolbarAdvancedFilteringComponent,
-    IgxGridToolbarComponent,
-    IgxGridToolbarExporterComponent,
-    IgxGridToolbarHidingComponent,
-    IgxGridToolbarPinningComponent,
-    IgxGridToolbarTitleComponent,
-    IgxGridToolbarDirective,
-    // IGX_GRID_EXCEL_STYLE_FILTER_DIRECTIVES:
-    IgxGridExcelStyleFilteringComponent,
-    IgxExcelStyleHeaderComponent,
-    IgxExcelStyleSortingComponent,
-    IgxExcelStylePinningComponent,
-    IgxExcelStyleHidingComponent,
-    IgxExcelStyleSelectingComponent,
-    IgxExcelStyleClearFiltersComponent,
-    IgxExcelStyleConditionalFilterComponent,
-    IgxExcelStyleMovingComponent,
-    IgxExcelStyleSearchComponent,
-    IgxExcelStyleColumnOperationsTemplateDirective,
-    IgxExcelStyleFilterOperationsTemplateDirective,
-    IgxExcelStyleLoadingValuesTemplateDirective,
-    // IGX_GRID_VALIDATION_DIRECTIVES:
-    IgxColumnRequiredValidatorDirective,
-    IgxColumnMinValidatorDirective,
-    IgxColumnMaxValidatorDirective,
-    IgxColumnEmailValidatorDirective,
-    IgxColumnMinLengthValidatorDirective,
-    IgxColumnMaxLengthValidatorDirective,
-    IgxColumnPatternValidatorDirective,
-    // IGX_PAGINATOR_DIRECTIVES:
-    IgxPaginatorComponent,
-    IgxPageNavigationComponent,
-    IgxPageSizeSelectorComponent,
-    IgxPaginatorContentDirective,
-    IgxPaginatorDirective
-] as const;
-
 export * from './hierarchical-grid.module';
